Integrated `backpack-react-scripts` in favour of ejected create-react-app
