+++ app/.eslintrc
@@ -1,0 +1,10 @@
+{
+  "extends": "skyscanner",
+  "env": {
+    "browser": true,
+    "jest": true
+  },
+  "rules": {
+    "import/no-extraneous-dependencies": [0]
+  }
+}
