+++ app/.gitignore
@@ -3,10 +3,13 @@
 # dependencies
 node_modules
 
+# testing
+coverage
+
 # production
 build
 
 # misc
 .DS_Store
+.env
 npm-debug.log
-Thumbs.db
