+++ app/README.md
@@ -1,0 +1,20 @@
+> *This project was bootstrapped with [Create React App][1] so please refer to [it's documentation][2] for more usage information.
+
+## npm script commands
+
+`backpack-react-scripts` is listed as a `devDependency` in your [package.json][3] and exposes the following
+npm scripts:
+
+- `npm start`: Runs the app in development mode. Open [http://localhost:3000/](http://localhost:3000/) to 
+  view it in the browser. The page will reload if you make edits. You will see the build errors in the console.
+- `npm run build`: Runs the test watcher in an interactive mode. By default, runs tests related to files changes 
+  since the last commit.
+- `npm test`: Builds the app for production to the build folder. It correctly bundles React in production mode 
+  and optimizes the build for the best performance. The build is minified and the filenames include the hashes.
+  Your app is ready to be deployed!
+- `npm run lint`: Lints all javascript using [`eslint-config-skyscanner`][4].
+
+[1]: https://github.com/facebookincubator/create-react-app
+[2]: https://github.com/facebookincubator/create-react-app/blob/master/packages/react-scripts/template/README.md
+[3]: ./package.json
+[4]: https://www.npmjs.com/package/eslint-config-skyscanner
