--- app/config/babel.dev.js
@@ -1,20 +1,0 @@
-module.exports = {
-  babelrc: false,
-  cacheDirectory: true,
-  presets: [
-    'babel-preset-es2015',
-    'babel-preset-es2016',
-    'babel-preset-react'
-  ].map(require.resolve),
-  plugins: [
-    'babel-plugin-syntax-trailing-function-commas',
-    'babel-plugin-transform-class-properties',
-    'babel-plugin-transform-object-rest-spread'
-  ].map(require.resolve).concat([
-    [require.resolve('babel-plugin-transform-runtime'), {
-      helpers: false,
-      polyfill: false,
-      regenerator: true
-    }]
-  ])
-};
