--- app/config/flow/css.js.flow
@@ -1,1 +1,0 @@
-// @flow
