+++ app/package.json
@@ -1,64 +1,22 @@
 {
   "name": "app",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "private": true,
   "devDependencies": {
-    "autoprefixer": "6.3.7",
-    "babel-core": "6.11.4",
-    "babel-eslint": "6.1.2",
-    "babel-loader": "6.2.4",
-    "babel-plugin-syntax-trailing-function-commas": "6.8.0",
-    "babel-plugin-transform-class-properties": "6.11.5",
-    "babel-plugin-transform-object-rest-spread": "6.8.0",
-    "babel-plugin-transform-react-constant-elements": "6.9.1",
-    "babel-plugin-transform-runtime": "6.12.0",
-    "babel-preset-es2015": "6.9.0",
-    "babel-preset-es2016": "6.11.3",
-    "babel-preset-react": "6.11.1",
-    "babel-runtime": "6.11.6",
-    "case-sensitive-paths-webpack-plugin": "1.1.2",
-    "chalk": "1.1.3",
-    "cross-spawn": "4.0.0",
-    "css-loader": "0.23.1",
-    "eslint": "3.1.1",
-    "eslint-loader": "1.4.1",
-    "eslint-plugin-flowtype": "2.4.0",
-    "eslint-plugin-import": "1.12.0",
-    "eslint-plugin-jsx-a11y": "2.0.1",
-    "eslint-plugin-react": "5.2.2",
-    "extract-text-webpack-plugin": "1.0.1",
-    "file-loader": "0.9.0",
-    "filesize": "3.3.0",
-    "fs-extra": "0.30.0",
-    "gzip-size": "3.0.0",
-    "html-webpack-plugin": "2.22.0",
-    "json-loader": "0.5.4",
-    "node-sass": "^3.8.0",
-    "opn": "4.0.2",
-    "postcss-loader": "0.9.1",
-    "promise": "7.1.1",
-    "raw-loader": "^0.5.1",
-    "rimraf": "2.5.4",
-    "sass-loader": "^4.0.0",
-    "style-loader": "0.13.1",
-    "url-loader": "0.5.7",
-    "webpack": "1.13.1",
-    "webpack-dev-server": "1.14.1",
-    "whatwg-fetch": "1.0.0"
-  },
-  "dependencies": {
-    "bpk-mixins": "^1.2.0",
-    "bpk-stylesheets": "0.0.18",
-    "bpk-tokens": "^8.2.0",
+    "backpack-react-scripts": "0.0.7",
+    "bpk-mixins": "^6.4.2",
+    "bpk-stylesheets": "^3.0.0",
     "lodash": "^4.15.0",
-    "react": "^15.2.1",
-    "react-dom": "^15.2.1"
+    "react": "^15.4.0",
+    "react-dom": "^15.4.0"
   },
+  "dependencies": {},
   "scripts": {
-    "start": "node ./scripts/start.js",
-    "build": "node ./scripts/build.js"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "lint": "eslint . --ignore-path .gitignore --ext .js,.jsx",
+    "eject": "react-scripts eject"
   },
-  "eslintConfig": {
-    "extends": "./config/eslint.js"
-  }
+  "proxy": "http://partners.api.skyscanner.net"
 }
