--- app/src/App.scss
@@ -1,1 +1,0 @@
-@import "~bpk-stylesheets/base.scss";
