+++ app/src/index.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import 'bpk-stylesheets/base';
+import 'bpk-stylesheets/base.css';
+
 import App from './App';
 
 ReactDOM.render(
-  <App />,
-  document.getElementById('root')
+  React.createElement(App),
+  document.getElementById('root'),
 );
